+++ src/containers/Posts/Posts.js
@@ -6,13 +6,14 @@ import Post from "components/Post/Post";
 import PostModal from 'components/PostModal/PostModal';
 import service from "api/service";
 import fbService from "api/fbService";
-import {actionTypes} from "context/actionTypes";
+//import {actionTypes} from "context/actionTypes";
+import {setReduxPosts,getMoreReduxPosts,hasMoreReduxPosts} from "actions/postActions";
 
 import load from "assets/load.gif";
 import noResults from "assets/noResults.jpg";
 
 import './Posts.scss';
-import { reduxActionTypes } from "reducers/reduxActionTypes";
+
 
 const limit = 8;
 
@@ -213,30 +214,15 @@ export class Posts extends Component {
 
 const mapStateToProps = (state)=>{
   return {
-    posts:state.posts,
-    hasMore:state.hasMore
+    posts:state.postsData.posts,
+    hasMore:state.postsData.hasMore
   }
 }
 
 const mapDispatchToProps ={
-  setReduxPosts:(posts)=>({
-    type:reduxActionTypes.SET_POSTS,
-    payload:{
-      posts,
-    }
-  }),
-  getMoreReduxPosts:(posts)=>({
-    type:reduxActionTypes.GET_MORE_POSTS,
-    payload:{
-      posts,
-    }
-  }),
-  hasMoreReduxPosts:(hasMore)=>({
-    type:reduxActionTypes.HAS_MORE_POSTS,
-    payload:{
-      hasMore,
-    }
-  })
+  setReduxPosts,
+  getMoreReduxPosts,
+  hasMoreReduxPosts
 }
 
 export default connect(mapStateToProps,mapDispatchToProps) (Posts);
