+++ src/reducers/store.js
@@ -1,42 +1,16 @@
-import {createStore} from 'redux';
-import { reduxActionTypes } from './reduxActionTypes';
+import {createStore,combineReducers} from 'redux';
+import postReducer from './postReducer';
+import countReducer from './countReducer';
 
-const initialState = {
-    count:0,
-    posts:null,
-    hasMore:true,
-  }
+// const initialState = {
+//     count:0,
+//     posts:null,
+//     hasMore:true,
+//   }
   
-  const reducer = (state = initialState,action)=>{
-    switch(action.type){
-        case reduxActionTypes.INCREMENT_COUNT:
-            return {
-                ...state,
-                count:state.count+1
-            }
-        case reduxActionTypes.DECREMENT_COUNT:
-                return {
-                ...state,
-                count:state.count-1
-            }
-        case reduxActionTypes.SET_POSTS:
-            return {
-                ...state,
-                posts:action.payload.posts
-            }    
-        case reduxActionTypes.GET_MORE_POSTS:
-            return {
-                ...state,
-                posts:[...state.posts, ...action.payload.posts]
-            }    
-        case reduxActionTypes.HAS_MORE_POSTS:
-            return {
-                ...state,
-                hasMore:action.payload.hasMore
-            }    
-        default:
-        return state;  
-    }
-  }
+  const reducers = combineReducers({
+      postsData: postReducer,
+      count: countReducer,
+  })
   
-  export const store = createStore(reducer);
+  export const store = createStore(reducers);
